+++ packages/cmrjs/src/cmr-utils.js
@@ -65,12 +65,13 @@ function getFileDescription(file, urlType = 'distribution') {
   return filename ? `Download ${filename}` : 'File to download';
 }
 
-const isECHO10File = (filename) => filename.endsWith('cmr.xml');
-const isUMMGFile = (filename) => filename.endsWith('cmr.json');
-const isISOFile = (filename) => filename.endsWith('cmr_iso.xml');
-const isCMRFilename = (filename) => isECHO10File(filename)
-  || isUMMGFile(filename)
-  || isISOFile(filename);
+const isECHO10Filename = (filename) => filename.endsWith('cmr.xml');
+const isUMMGFilename = (filename) => filename.endsWith('cmr.json');
+const isISOFilename = (filename) => filename.endsWith('.iso.xml');
+const isCMRISOFilename = (filename) => filename.endsWith('cmr_iso.xml');
+const isCMRFilename = (filename) => isECHO10Filename(filename)
+  || isUMMGFilename(filename)
+  || isCMRISOFilename(filename);
 
 const constructCmrConceptLink = (conceptId, extension) => `${getSearchUrl()}concepts/${conceptId}.${extension}`;
 
@@ -86,18 +87,31 @@ function isCMRFile(fileobject) {
 }
 
 /**
+ * Returns True if this object can be determined to be an ISO file object.
+ *
+ * @param {Object} fileobject
+ * @returns {boolean} true if object references an ISO file metadata.
+ */
+function isISOFile(fileobject) {
+  const filename = fileobject.key || fileobject.name || fileobject.filename || '';
+  return isISOFilename(filename) || isCMRISOFilename(filename);
+}
+
+/**
  * Extracts CMR file objects from the specified granule object.
  *
  * @param {Object} granule - granule object containing CMR files within its
  *    `files` property
  * @param {Array<Object>} granule.files - array of files for a granule
  * @param {string} granule.granuleId - granule ID
+ * @param {Function} filterFunc - function to determine if the given file object is a
+      CMR file; defaults to `isCMRFile`
  * @returns {Array<Object>} an array of CMR file objects, each with properties
  *    `granuleId`, `filename`, and possibly `etag` (if present on input)
  */
-function granuleToCmrFileObject({ granuleId, files = [] }) {
+function granuleToCmrFileObject({ granuleId, files = [] }, filterFunc = isCMRFile) {
   return files
-    .filter(isCMRFile)
+    .filter(filterFunc)
     .map((file) => ({
       // Include etag only if file has one
       ...pick(file, 'etag'),
@@ -111,11 +125,13 @@ function granuleToCmrFileObject({ granuleId, files = [] }) {
  * Reduce granule object array to CMR files array
  *
  * @param {Array<Object>} granules - granule objects array
+ * @param {Function} filterFunc - function to determine if the given file object is a
+      CMR file; defaults to `isCMRFile`
  *
  * @returns {Array<Object>} - CMR file object array: { filename, granuleId }
  */
-function granulesToCmrFileObjects(granules) {
-  return granules.flatMap(granuleToCmrFileObject);
+function granulesToCmrFileObjects(granules, filterFunc = isCMRFile) {
+  return granules.flatMap((granule) => granuleToCmrFileObject(granule, filterFunc));
 }
 
 /**
@@ -202,10 +218,10 @@ async function publish2CMR(cmrPublishObject, creds, cmrRevisionId) {
   const cmrClient = new CMR(creds);
 
   // choose xml or json and do the things.
-  if (isECHO10File(cmrPublishObject.filename)) {
+  if (isECHO10Filename(cmrPublishObject.filename)) {
     return await publishECHO10XML2CMR(cmrPublishObject, cmrClient, cmrRevisionId);
   }
-  if (isUMMGFile(cmrPublishObject.filename)) {
+  if (isUMMGFilename(cmrPublishObject.filename)) {
     return await publishUMMGJSON2CMR(cmrPublishObject, cmrClient, cmrRevisionId);
   }
 
@@ -290,14 +306,14 @@ const metadataObjectFromCMRXMLFile = (cmrFilename, etag) =>
  * @returns {Promise<Object>} metadata object from the file
  * @throws {Error} if the specified filename does not represent an ECHO-10 XML
  *    file or a UMMG file
- * @see isECHO10File
- * @see isUMMGFile
+ * @see isECHO10Filename
+ * @see isUMMGFilename
  */
 function metadataObjectFromCMRFile(cmrFilename, etag) {
-  if (isECHO10File(cmrFilename)) {
+  if (isECHO10Filename(cmrFilename) || isISOFilename(cmrFilename)) {
     return metadataObjectFromCMRXMLFile(cmrFilename, etag);
   }
-  if (isUMMGFile(cmrFilename)) {
+  if (isUMMGFilename(cmrFilename)) {
     return metadataObjectFromCMRJSONFile(cmrFilename, etag);
   }
   throw new Error(
@@ -885,9 +901,9 @@ async function updateCMRMetadata({
   let metadataObject;
   let etag;
 
-  if (isECHO10File(filename)) {
+  if (isECHO10Filename(filename)) {
     ({ metadataObject, etag } = await updateEcho10XMLMetadata(params));
-  } else if (isUMMGFile(filename)) {
+  } else if (isUMMGFilename(filename)) {
     ({ metadataObject, etag } = await updateUMMGMetadata(params));
   } else {
     throw new errors.CMRMetaFileNotFound(`Invalid CMR filetype: ${filename}`);
@@ -1024,7 +1040,7 @@ async function getGranuleTemporalInfo(granule) {
 
   const cmrFilename = cmrFile[0].filename;
 
-  if (isISOFile(cmrFilename)) {
+  if (isCMRISOFilename(cmrFilename)) {
     const metadata = await metadataObjectFromCMRXMLFile(cmrFilename);
     const metadataMI = metadata['gmd:DS_Series']['gmd:composedOf']['gmd:DS_DataSet']['gmd:has']['gmi:MI_Metadata'];
 
@@ -1044,7 +1060,7 @@ async function getGranuleTemporalInfo(granule) {
 
     return { beginningDateTime, endingDateTime, productionDateTime, lastUpdateDateTime };
   }
-  if (isECHO10File(cmrFilename)) {
+  if (isECHO10Filename(cmrFilename)) {
     const metadata = await metadataObjectFromCMRXMLFile(cmrFilename);
     const beginningDateTime = get(metadata.Granule, 'Temporal.RangeDateTime.BeginningDateTime');
     const endingDateTime = get(metadata.Granule, 'Temporal.RangeDateTime.EndingDateTime');
@@ -1054,7 +1070,7 @@ async function getGranuleTemporalInfo(granule) {
       beginningDateTime, endingDateTime, productionDateTime, lastUpdateDateTime,
     };
   }
-  if (isUMMGFile(cmrFilename)) {
+  if (isUMMGFilename(cmrFilename)) {
     const metadata = await metadataObjectFromCMRJSONFile(cmrFilename);
     const beginningDateTime = get(metadata, 'TemporalExtent.RangeDateTime.BeginningDateTime');
     const endingDateTime = get(metadata, 'TemporalExtent.RangeDateTime.EndingDateTime');
@@ -1087,9 +1103,11 @@ module.exports = {
   granulesToCmrFileObjects,
   isCMRFile,
   isCMRFilename,
-  isECHO10File,
+  isCMRISOFilename,
+  isECHO10Filename,
   isISOFile,
-  isUMMGFile,
+  isISOFilename,
+  isUMMGFilename,
   metadataObjectFromCMRFile,
   publish2CMR,
   reconcileCMRMetadata,